requirements.txt
#pynamodb
#boto3
#botocore
fastapi
pydantic[email]
python-multipart
aiofiles
jinja2
pymongo
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pytest
pytest-asyncio
async_asgi_testclient
requests
fastapi-crudrouter
fastapi-permissions
fastapi-admin
fastapi-users[tortoise-orm]
aiosqlite
asyncpg
aiomysql
asynctest
